+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+### Added
+- Currency conversion. (#1)
 
 ## [1.1.0] - 2017-09-19
 ### Added